requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.11.0
beanie==1.27.0
click==8.1.7
dnspython==2.7.0
exceptiongroup==1.2.2
#fastapi==0.115.5
fastapi
h11==0.14.0
idna==3.10
lazy-model==0.2.0
loguru==0.7.2
motor==3.6.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
# pydantic==2.10.2
# pydantic==1.10.14
# pydantic_core==2.27.1
pylnk3==0.4.2
pymongo==4.9.2
python-calamine==0.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uvicorn==0.32.1
sqlmodel
asyncpg
python-multipart